MirLua: code optimization
git-svn-id: http://svn.miranda-ng.org/main/trunk@16098 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/MirLua/src/mlua_metatable.h
@@ -41,7 +41,6 @@ struct MTField : public MZeroedObject
 			return res;
 		}
 	}
-
 };
 
 template<typename T>
@@ -100,7 +99,7 @@ private:
 			lua_pushboolean(L, field->GetValue<BOOL>(obj));
 			break;
 		case LUA_TINTEGER:
-			lua_pushinteger(L, field->GetValue<long>(obj));
+			lua_pushinteger(L, field->GetValue<long long>(obj));
 			break;
 		case LUA_TNUMBER:
 			lua_pushnumber(L, field->GetValue<intptr_t>(obj));
